requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
distlib==0.3.6
Django==4.2
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.2.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.11.0
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
phonenumbers==8.13.23
Pillow==10.0.0
platformdirs==3.2.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
twilio==8.11.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
yarl==1.9.4
zope.interface==6.1
